# Eliminate locking in (anti)affinity calculations